[Refactor]Environment Object 제거 후 StateObject로 변환
AsyncswiftApp의 Environment Object 생성 -> MantabView의 StateObject로 변경
