+++ AsyncSwift/AsyncSwiftApp.swift
@@ -10,21 +10,10 @@ import SwiftUI
 @main
 struct AsyncSwiftApp: App {
     @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
-    @ObservedObject var appData: AppData = AppData()
     
     var body: some Scene {
         WindowGroup {
             MainTabView()
-                .environmentObject(appData)
-                .onOpenURL { url in
-                    Task {
-                        if await appData.checkLink(url: url) {
-                            print("Success Link URL: \(url)")
-                        } else {
-                            print("Fail Link URL: \(url)")
-                        }
-                    }
-                }
         }
     }
 }
